requirements.txt
roboflow>=1.1.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
numpy>=1.24.0

